[PERF] flatten DS.Model
